requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
pillow==10.4.0
pytesseract==0.3.10
pdf2image==1.17.0
python-dotenv==1.0.1
schedule==1.2.1
pymongo==4.8.0
bcrypt==4.2.0
scikit-learn==1.5.1
streamlit-cookies-manager==0.2.0
